requirements.txt
opencc-python-reimplemented>=0.1.7
streamlit>=1.44.0
python-docx>=1.1.2
docx2txt>=0.9
